% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inplace_ops.R
\name{\%-=\%}
\alias{\%-=\%}
\title{Minus and assign}
\usage{
t \%-=\% m
}
\arguments{
\item{t}{A number.}

\item{m}{Another number.}
}
\value{
The result of x - y.
}
\description{
A useful function maybe.
}
