% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inplace_ops.R
\name{\%+=\%}
\alias{\%+=\%}
\title{Add and assign}
\usage{
t \%+=\% s
}
\arguments{
\item{t}{Target variable (numeric).}

\item{s}{Increment (numeric).}
}
\value{
The result of adding \code{s} to \code{t} and assigning it to \code{t}.
}
\description{
This operator adds the right-hand side value to the left-hand side variable and assigns the result to the left-hand side variable.
}
\examples{
x <- 1
x \%+=\% 10
x
}
