% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{inplaceops-package}
\alias{inplaceops}
\alias{inplaceops-package}
\title{inplaceops: In-Place Operations}
\description{
In-place operations
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/a-klier/r_package_demo}
  \item Report bugs at \url{...}
}

}
\author{
\strong{Maintainer}: Anne Klier \email{anne.klier@iis.fraunhofer.de}

}
