% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiply.R
\name{multiply}
\alias{multiply}
\title{Multiply two numbers (vectorised)}
\usage{
multiply(a, b)
}
\arguments{
\item{a, b}{Numeric vectors.}
}
\value{
A numeric vector with \code{a * b}.
}
\description{
Simple, fast multiplication of two numeric vectors (element-wise). Recycles
following R's standard recycling rules.
}
\examples{
multiply(2, 3)          # 6
multiply(1:3, 10)       # 10 20 30
multiply(c(0.5, 2), 4)  # 2 8

}
