% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_function.R
\name{print_string_n_times}
\alias{print_string_n_times}
\title{Print text x times}
\usage{
print_string_n_times(text, n)
}
\arguments{
\item{text}{A string to be printed.}

\item{n}{An integer specifying how many times to print the text.}
}
\value{
\code{string}
}
\description{
\code{print_string_n_times()} prints text as often as specified.
}
\examples{
print_string_n_times("Hello, World!", 3)


}
