% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{zhengtst-package}
\alias{zhengtst}
\alias{zhengtst-package}
\title{zhengtst: A test package}
\description{
Just a test package.
}
\seealso{
Useful links:
\itemize{
  \item \url{NULL}
  \item Report bugs at \url{https://github.com/zhenglukai/zhengtst/issues}
}

}
\author{
\strong{Maintainer}: Lukai Zheng \email{zheng.lukai@campus.lmu.de}

}
