% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inplace_ops.R
\name{\%+=\%}
\alias{\%+=\%}
\title{In-Place Addition}
\usage{
t \%+=\% s
}
\arguments{
\item{t}{(\code{numeric})\cr
Variable to increase}

\item{s}{(\code{numeric})\cr
Amount to increase by}
}
\description{
Increase t by s
}
